extract new file
